Oberoi Pokhran Road 2 Thane
Oberoi Pokhran Road 2 is an awesome project by Oberoi Realty. This venture is known for its modern design and detailed floor planning. The residential units have been designed well and fantastically to provide a great relief zone to its buyers. The location of this venture is unbeatable and provides ease of access and comfort to people all around.
Oberoi Realty Flats comes in the configuration of 2, 3 and 4 BHK units. The residential units come in various carpet sizes and 
 are planned with a maximum area utilization policy letting in ample light and air flowing through wide balconies, and windows and revelling in the beautiful lush green environment. The interiors are carried out with smart finishing all around. The state-of-the-art amenities of this venture include a meditation centre, kids’ play area, swimming pool, gymnasium, landscaped garden, 80% open space, and a multipurpose hall. There are a host of fitness and leisure zones within the complex. It is a great funding project with high future returns.
This site is located in Thane at Pokhran Rd no. 2 . The neighbourhood of this property holds the best of the schools, hospitals, colleges, parks, gardens, restaurants, shopping malls, multiplexes, etc. This location also has remarkable connectivity to all essential and major hubs & places in Thane and Mumbai via Eastern Express Highway and Pokhran Road No 1.
Established in 1998, Oberoi Realty is one of India's premier companies engaged in real estate planning and development. With rich enterprise experience and a customer-oriented strategy with a commitment to quality and on-time delivery of its all projects, the group has carved a niche for itself in the realty sector. The vast portfolio comprising many projects across many cities, including Mumbai, Pune and Bangalore exhibits the unmatched expertise of the Group in the real property market. 
To know more - https://bit.ly/3Yw9569